Move integration test to new test class
